(PUP-2858) Deprecate the use of Puppet[:environment] == current
This changes defaults.rb to deprecate the use of :environment = :current
The future evaluator is always picked by the Parser::Factory and
the setting has no effect. It may have effect in future versions of
Puppet (for additional experiments when Puppet 4.0 is released). For
that purpose, the settings is not completely removed.