(PUP-2858) Remove the 3_1 checker and its factory
This removes the Checker_3_1 and its factory from the code base.
This hybrid checker contained less validation (less string) than
the 4.0 checker and its main purpose was to fail expressions that
could not be evaluated with the 3x AST based evaluation (such as -=).

It is also not meaningful to continue its development as the new
4_0 checker is far better. If there is a later need to check
for "backwards compatibility" to 3x, it is better to start fresh with a
new copy of the 4_0 evaluator.