(PUP-480) Change AccessOperator to not consider :undef from evaluator
This changes the AccessOperator logic to not consider :undef being
prouced by the evaluator (as key). It continues to support a hash with a
value stored under the key :undef and making that return that
value is accessed as $hsh[undef]. It is expected that hashes
follow the 4x convention (nil), but this attempts to also work
if the hash was constructed with 3x logic and uses :undef instead of
nil. (nil has priority if there is both a nil and a :undef key in
the same hash)