(PUP-1283) Relax win32-service constraint
Previously we were requiring an exact version match for the win32-service
gem. However, the ffi version of the gem has a regression and we expect a
new version of the gem to be released soon.

This commit relaxes the dependency so that we don't have to revisit this
issue in the near future.