(PUP-2532) Correct lookup of default parameters
The lookup of resource parameters via the type e.g. Notify[id][message]
performed the lookup in the calling scope, not the resource's closure
scope which resulted in parameter default values were being seen from
the perspective of the caller (they may be very different).

This corrects the lookup to use the resource's closure scope.