(PUP-2521) Remove reference to windows-pr ERROR_FILE_NOT_FOUND constant
Previously, bundle exec rspec spec/unit/provider/package/windows/package_spec.rb
would fail on windows with ruby 2 x64, with the error message:

    NameError: uninitialized constant Windows::Error

as it referenced a constant from windows-pr, which we are in the process of
removing.

This commit changes the code to use the same constant defined in
Puppet::Util::Windows::Error.