(maint) Error message is different under Windows 2012r2
Previously, the test was expecting a failed logon to raise an error
whose message contained:

    unknown user name or bad password.

This is correct in 2008r2 and earlier, but it appears Windows 2012
changed the error message to be:

    The user name or password is incorrect.

This commit updates the test to not be sensitive to the message
text. Instead we verify that the exception contains the correct Win32
error code (ERROR_LOGON_FAILURE).