(maint) Mark eventlog tests as conditionally pending on ruby 2
Previously, the tests would fail on ruby2, because under ruby 2 we are using
a newer win32-eventlog gem, whose constants are in a different namespace than
the version of the gem we are using for ruby 1.9.

This commit marks the tests as conditionally pending when running under Windows
and ruby2. The tests will be re-enabled when PUP-2889 is implemented.