(maint) Fix test to check desired functionality
This test had a flaw in it before where if the collection was not even
done, the test would still pass. In order to check that collection
applies to realized resources the resources need to be changed. This
adds an override so that we can see the effect of the collection
expression on realized resources.