(PUP-744) Create an HTTP site value object
This commit creates a site value object containing a scheme, host, and
port. A site identifies an HTTP(S) connection. It does not currently
take into account proxy host, port, user, password, as these are global
puppet settings.

The Site class defines the `eql?` and `hash` methods making it suitable
for use as the key in a hash table.