(PUP-2858) Remove evaluator setting
The evaluator selection is no longer possible and so the setting is not
needed. Although removing this setting could be considered to be
backwards incompatible, I don't think it should be considered such. The
issue is that puppet can no longer support the "current" value for the
evaluator and so anything we do will either behave differently (use the
future evaluator instead) or fail (remove the setting or cause it to
fail if the value is "current"). Since this is part of an experimental
feature, I think that simply removing it is a reasonable course of
action.