(PE-2912) Error on symlink in module install
Prior to this commit when encountering a symlink in a module file during install,
the install would error out with No such file or directory - /path/to/symlink.
This is confusing and doesn't show what the problem actually is.
This commit modifies the installer behavior to check for any symlinks after the
module is extracted, give errors for each symlink it finds, and die. This will
hopefully help module authors track down the actual reason for the failed install
more easily.