(maint) Pin rest-client gem to 1.6.7
This commit addresses spec failures under ruby 1.8.7.

Background:
Puppet uses the couchrest gem for the soon-to-be-deprecated
couchdb facts terminus. The couchrest gem has a pessimistic
dependency on the 1.6.x of the rest-client gem. The previous
release of the rest-client gem was 1.6.7 in 2011, but on July
8th 1.6.8 was released, and it introduced a dependency on
the rdoc 4 gem, which breaks various puppet code.

Since we're planning to deprecate this code anyway, and remove
it in puppet-4, I opted for the simple approach of pinning
this one dependency.