(maint) Remove the unused class Visitor::Visitor2
The Visitor2 class had an alternative method for making
polymorphic dispatch based on fact that calling a bound method
is faster than a send. The gain is however lost since bound methods
are instance specific and it is thus not possible to cache across
all instances. Using the Visitor2 also requires one extra step of setup.

Now that we have benchmarks, there was no improvement when using
Visitor2 over the regular Visitor, in fact there was a tendency for
Visitor2 to be slower (by ~1% in a couple of the runs).

This was expected, as polymorphic dispatching did not show up on the hot
list.