(PUP-2860) Remove future parser's old lexer
The old (intermediate) lexer was still hanging around even if it was
not being used. Cycles were also wasted on unit testing it.

The new lexer "Lexer2" has been in use for quite some time and the
old lexer is not useful even as a reference at this point.

This speeds up both loading and unit testing.