(perf) freeze all string tokens at the source (in the lexer)
This freezes all strings that are returned as tokens (varname,
sq and dq string, name, classref). Keywords were already frozen.

This optimizations means that these strings does not have to be frozen
later, and the values are faster to use as hash lookup without incurring
the penalty of again freezing (~10us per freeze call).

+ it is a sane thing to do anyway to avoid someone modifying the
strings.