(PUP-744) Remove monkey patching of Net::HTTP#ca_file=
Commit 553b2ad8a (circa 2007) monkey patched Net::HTTP to set
`Net::HTTP#ca_file=`. This method is needed to specify the file
containing the set of CA cert(s) that the SSL client will use to
authenticate SSL servers.

Note the `ca_file=` method is available in all versions of ruby 1.8.7
and up[1], so it is no longer necessary to monkey patch this.

[1] https://github.com/ruby/ruby/blob/v1_8_7/lib/net/https.rb#L145