(PUP-2936) Fix checksum during PMT build
Prior to this commit `puppet module build` was checksumming the source directory
and putting that in checksums.json.
This commit corrects that behavior so it checksums the built module instead.
Additionally, it removes the deprecated and confusing checksums field from
metadata.json because there is no reason to keep it around during a build.