(CPR-37)(PUP-...) vendor json_pure in OSX package
Per CPR-37, prior to this commit the dmg/pkg of puppet we shipped did not work
out of the box in OSX < 10.9 because of a dependency on json that is not
satisfied by system ruby. In OSX 10.9 system ruby is 2.0...prior to this it was
1.8.7 or less. This commit introduces a pre-automation task that retrieves the
json_pure 1.8.1 release from github and vendors it within the lib. This is the
approach we took with facter 2 and cfpropertylist for FACT-332 in aa9ee28b of
facter. This commit will drop json.rb and its lib dir in /Library/Ruby/Site
alongside puppet.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>