(PUP-744) Remove private Connection#connection method
Previously, the `Connection#connection` method was used in
`connection_spec.rb` to validate the state of the Net::HTTP object
created by the puppet connection.

Since the puppet connection no longer owns an Net::HTTP object (it may
use newly created or cached instances), the `#connection` method
doesn't make sense, and since it was private, can be removed.

The tests that relied on the private method have been updated, and in
some cases moved to the factory_spec, e.g. verifying proxy settings.