(PUP-2936) Add warning for 'checksums'
This commit adds a warning to `puppet module build` when it finds a 'checksums'
field in metadata.json. This was done to make people aware of the fact the field
is deprecated and will not be used by the builder, although the installer/upgrader
and other module tools will use the field if checksums.json is not present.