(maint) Confine enc_provides_node_when_store_configs_enabled to debian
This test has a long history of being weaned from troublesome platforms
where there are issues getting the packages installed.  The latest problem
child is centos5, now that epel is not enabled on our vcloud templates
by default.  So I'm just confining the test here to run only on debian
(except lucid, which is another problem child for this test).  Since
activerecord is slated to be removed as a store_configs backend in Puppet
4, this test is on the chopping block soon regardless.