(PUP-744) Fix off-by-one redirect limit error
Previously, the `redirect_limit` specified the maximum number of HTTP
requests that the connection would make, not the maximum number of
redirects that it would follow. As a result, a limit of 0 would prevent
any HTTP requests from being made. This was introduced in 60f22eb1 when
the redirect logic was added.

This commit changes the limit so that 0 means don't follow redirects,
1 means follow one redirect, etc.