(maint) Fix spec order issue for Travis CI.
Because of the order Travis runs specs, some environment caching issues
were found that caused some specs to fail because they made the
assumption that the production environment wasn't already in the cache.
Thus, they relied on changing Puppet[:modulepath] to have an effect.

The fix is to override the environments for these two
specs and return an environment we know has the expected module path.