(maint) Update comment about request_* methods
The puppet connection class has request_get, etc methods that accept a
block. I tried to deprecate them, to eliminate the possibility that a
caller could borrow a connection from the pool indefinitely.

However, I found that we have to pass a block to ruby's request method
in order to stream the response body. If you don't pass a block, ruby
will always read the entire response body into memory. If you later
try to stream the body:

    response.read_body { |chunk| yield uncompressor.uncompress(chunk) } }

It will fail with:

    Net::HTTPOK#read_body called twice

This commit updates the comment so it's clear why these other
request_methods exist.