(PUP-2860) Fix comments as pointed out in review
Fix files with same comment after splitting models. The implementation
now points out that it uses the meta part and mixes in the
implementation.

The possible optimization logged in a ticket, and ticket is instead
referenced in the comments.

Manual inlining of accessor commented to avoid it being DRYed out.