(PUP-2898) Make Resource Expr comply with spec (left QualifiedReference)
The LHS (type name) of a Resource Expression is specified as
QualifiedName or QualifiedReference (or to be a Resource[T]). The
implementation did not work when given a QualifiedReference (runtime
error, due to errors in the error checking).

This both fixes the runtime error, and makes a string that
complies with QualifiedReference work. (An actual QualifiedReference
evaluates to Resource[T], so that already worked).

This also removes a call to debugger in relationship operator (that
would be triggered if user tried to create a resource from a non
CatalogEntry type).