(PUP-2991) Restore ability to manage shells on solaris
Background: for PUP-1448 commit 73b0c2b0 added a new feature
for user providers `manage_shells`. The intent of that commit
was to add this feature to all existing user providers which
had that feature, and several were covered, but Solaris was
missed.

This commit simply adds the `manage_shells` feature to the
Solaris user provider, following up on the intent in PUP-1448.